+++ standalone-scripts/payment-banner-hider/src/index.ts
@@ -19,7 +19,7 @@ import {
     type PaymentBannerHiderApi,
 } from "./types";
 
-const VERSION = "3.100.0";
+const VERSION = "3.101.0";
 
 export class PaymentBannerHider implements PaymentBannerHiderApi {
     public readonly version = VERSION;